# docs/swagger.yaml
basePath: /
definitions:
...
  models.Webhook:
    description: Webhook configuration for event notifications
    properties:
      created_at:
        type: string
      created_by:
        example: 507f1f77bcf86cd799439011
        type: string
      description:
        example: Sync contacts to remote CRM
        type: string
      events:
        example:
        - record.updated
        - record.created
        items:
          type: string
        type: array
      headers:
        additionalProperties:
          type: string
        description: Custom headers to send
        type: object
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      is_active:
        example: true
        type: boolean
      module_name:
        description: 'Optional: limit to specific module'
        example: leads
        type: string
      secret:
        description: For HMCA signature
        example: my_secret_key
        type: string
      updated_at:
        type: string
      url:
        example: https://example.com/webhook
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: This is a sample server using Fiber, Uber Fx, and GORM.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Microservice Demo API
  version: "1.0"
paths:
...
  /api/webhooks:
    get:
      description: Get a list of all webhooks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Webhook'
            type: array
      summary: List all webhooks
      tags:
      - webhooks
    post:
      consumes:
      - application/json
      description: Register a new webhook url
      parameters:
      - description: Webhook Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Webhook'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create a new webhook
      tags:
      - webhooks
  /api/webhooks/{id}:
    delete:
      description: Delete a webhook by ID
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Delete a webhook
      tags:
      - webhooks
    get:
      description: Get a webhook by ID
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Webhook'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get a webhook
      tags:
      - webhooks
    put:
      consumes:
      - application/json
      description: Update a webhook by ID
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Update a webhook
      tags:
      - webhooks
...
swagger: "2.0"